Add attachment to user
