+++ app/models/invoice.rb
@@ -1,4 +1,5 @@
 class Invoice < ApplicationRecord
+  include Attachment
   include CustomFields
   include LocalizeInput
 
@@ -10,28 +11,12 @@ class Invoice < ApplicationRecord
 
   validates_presence_of :supplier_id
   validates_numericality_of :amount, :deposit, :deposit_credit
-  validate :valid_attachment
 
   scope :unpaid, -> { where(paid_on: nil) }
   scope :without_financial_link, -> { where(financial_link: nil) }
 
-  attr_accessor :delete_attachment
-
   # Replace numeric seperator with database format
   localize_input_of :amount, :deposit, :deposit_credit
-
-  def attachment=(incoming_file)
-    self.attachment_data = incoming_file.read
-    # allow to soft-fail when FileMagic isn't present and removed from Gemfile (e.g. Heroku)
-    self.attachment_mime = defined?(FileMagic) ? FileMagic.new(FileMagic::MAGIC_MIME).buffer(self.attachment_data) : 'application/octet-stream'
-  end
-
-  def delete_attachment=(value)
-    if value == '1'
-      self.attachment_data = nil
-      self.attachment_mime = nil
-    end
-  end
 
   def user_can_edit?(user)
     user.role_finance? || (user.role_invoices? && !self.paid_on && self.created_by.try(:id) == user.id)
@@ -58,16 +43,5 @@ class Invoice < ApplicationRecord
     return net_amount unless orders.any?
 
     orders_sum + orders_transport_sum
-  end
-
-  protected
-
-  def valid_attachment
-    if attachment_data
-      mime = MIME::Type.simplified(attachment_mime)
-      unless ['application/pdf', 'image/jpeg'].include? mime
-        errors.add :attachment, I18n.t('model.invoice.invalid_mime', :mime => mime)
-      end
-    end
   end
 end
