+++ app/models/user.rb
@@ -1,6 +1,7 @@
 require 'digest/sha1'
 # specific user rights through memberships (see Group)
 class User < ApplicationRecord
+  include Attachment
   include CustomFields
   # TODO: acts_as_paraniod ??
