+++ config/locales/nl.yml
@@ -198,6 +198,8 @@ nl:
         user_list: Verantwoordelijken
         workgroup: Werkgroep
       user:
+        attachment: Bijlage
+        delete_attachment: Verwijder bijlage
         email: Email
         first_name: Voornaam
         iban: IBAN
@@ -363,6 +365,7 @@ nl:
       edit:
         title: Lid bewerken
       form:
+        attachment_hint: Alleen JPEG en PDF zijn toegestaan.
         create_ordergroup: Maak een huishouden met dezelfde naam en voeg een gebruiker toe.
         send_welcome_mail: Verstuur een welkomstmail naar de gebruiker.
       index:
@@ -1370,8 +1373,7 @@ nl:
       no_delete_last: Er moet ten minste één financiëel transactie type bestaan.
     group_order:
       stock_ordergroup_name: Voorraad (%{user})
-    invoice:
-      invalid_mime: heeft ongeldig MIME type (%{mime})
+    invalid_mime: heeft ongeldig MIME type (%{mime})
     membership:
       no_admin_delete: Lidmaatschap kan niet beeindigd worden. Je bent de laatste administrator.
     order_article:
